requirements.txt
numpy
scipy
pandas
matplotlib
progressbar2
ruamel.yaml
pylint
ipython
